+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 0.0.6 (2019-04-17)
 
 * enhancement: aesthetics
 * bugfix: buttons should not submit the form
+* bugfix: rwVal should be pushed into rwAuto to affect the graph, and not "live"
 
 
 Version 0.0.5 (2019-04-17)